Refactor recharacterize feature: split service, dedupe templates
End-to-end cleanup of the recharacterize tool after its incremental build-out.
Behavior is unchanged; this addresses layering, duplication, and navigability.

Backend:
- Split the 941-line recharacterize_services.py into a same-named package
  (constants / resolution / evaluation / preview / apply / llm) with __init__
  re-exports, so all call sites and tests import exactly as before.
- Promote gemini_services._loads_gemini_json to a public loads_gemini_json;
  stop reaching into another module's private helper.
- Make RecharacterizeOperationForm take a required keyword-only `catalogs`,
  fully severing the Forms->Services dependency; the view builds catalogs once
  per request and threads them through _render_main / _prefilled_form.
- Consolidate op-index parsing on the bounds-checked _valid_op_index; push the
  "invalid index -> nothing" rule into _evaluate_at so the views pass the index
  straight through.

Frontend:
- Extract the copy-pasted inline-confirm control into _inline-confirm-button.html
  (Apply + Revert), unify the four error surfaces into _alert.html, and factor
  the revise-with-agent form into its own partial.
- Replace emoji icons with Bootstrap Icons; remove the stale ID-coupled CSS rule
  and the inline display:none in favor of a reusable .hidden utility.

Full test suite passes.
