+++ ledger/templates/api/components/recharacterize-history.html
@@ -20,25 +20,8 @@
                     {% if change.is_reverted %}
                     <span class="muted">Reverted {{ change.reverted_at|date:"M j, Y g:i A" }}</span>
                     {% else %}
-                    <span x-data="{ confirming: false }">
-                        <button type="button" class="btn btn-danger"
-                                x-show="!confirming"
-                                @click="confirming = true">
-                            Revert
-                        </button>
-                        <span class="inline-confirm" x-show="confirming" x-cloak>
-                            <button type="button" class="btn btn-danger"
-                                    hx-post="{% url 'recharacterize-revert' %}?change={{ change.id }}"
-                                    hx-target="#recharacterize-main"
-                                    hx-swap="outerHTML">
-                                Confirm
-                            </button>
-                            <button type="button" class="btn btn-secondary"
-                                    @click="confirming = false">
-                                Cancel
-                            </button>
-                        </span>
-                    </span>
+                    {% url 'recharacterize-revert' as revert_url %}
+                    {% include "api/components/_inline-confirm-button.html" with trigger_label="Revert" confirm_url=revert_url confirm_param="change" confirm_value=change.id %}
                     {% endif %}
                 </div>
             </div>
