+++ ledger/templates/api/components/recharacterize-main.html
@@ -32,19 +32,19 @@ whole session (chat + plan), so it lives here — global to both tabs.
                 {% endfor %}
                 <div id="chat-thinking" class="chat-msg chat-thinking">
                     <span class="chat-role">Agent</span>
-                    <div class="chat-text">⚙️ Thinking…</div>
+                    <div class="chat-text"><i class="bi bi-hourglass-split"></i> Thinking…</div>
                 </div>
             </div>
 
             {% if error %}
-            <div class="alert alert-danger mt-3" role="alert">
-                <div title="{{ error.detail }}">❌ {{ error.label }} — {{ error.message }}</div>
+            <div class="mt-3">
+                {% include "api/components/_alert.html" with icon="exclamation-triangle" label=error.label message=error.message detail=error.detail %}
                 <button type="button" class="btn btn-secondary btn-sm mt-2"
                         hx-post="{% url 'recharacterize-retry' %}"
                         hx-target="#recharacterize-main"
                         hx-swap="outerHTML"
                         hx-indicator="#chat-thinking"
-                        hx-disabled-elt="this">↻ Retry</button>
+                        hx-disabled-elt="this"><i class="bi bi-arrow-clockwise"></i> Retry</button>
             </div>
             {% endif %}
